+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - add NIP-32 (labeling)
 - add NIP-38 (user statuses)
 - add NIP-03 (OpenTimestamps attestations)
+- add NIP-37 (draft events)
 
 ## v0.1.0 (2023-06-07)